1.Extract common function to toggle environment type 2.Set default folder environment type to be key-value pair
