--- packages/insomnia/src/ui/components/editors/environment-utils.ts
@@ -1,46 +1,0 @@
-import { NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME } from '../../../templating';
-
-// NeDB field names cannot begin with '$' or contain a period '.'
-// Docs: https://github.com/DeNA/nedb#inserting-documents
-const INVALID_NEDB_KEY_REGEX = /^\$|\./;
-
-export const ensureKeyIsValid = (key: string, isRoot: boolean): string | null => {
-  if (key.match(INVALID_NEDB_KEY_REGEX)) {
-    return `"${key}" cannot begin with '$' or contain a '.'`;
-  }
-
-  if (key === NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME && isRoot) {
-    return `"${NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME}" is a reserved key`;
-  }
-
-  return null;
-};
-
-/**
- * Recursively check nested keys in and immediately return when an invalid key found
- */
-export function checkNestedKeys(obj: Record<string, any>, isRoot = true): string | null {
-  for (const key in obj) {
-    let result: string | null = null;
-
-    // Check current key
-    result = ensureKeyIsValid(key, isRoot);
-
-    // Exit if necessary
-    if (result) {
-      return result;
-    }
-
-    // Check nested keys
-    if (typeof obj[key] === 'object') {
-      result = checkNestedKeys(obj[key], false);
-    }
-
-    // Exit if necessary
-    if (result) {
-      return result;
-    }
-  }
-
-  return null;
-}
