+++ packages/insomnia/src/ui/routes/environments.tsx
@@ -6,18 +6,17 @@ import { NavLink, useFetcher, useParams, useRouteLoaderData } from 'react-router
 
 import { DEFAULT_SIDEBAR_SIZE } from '../../common/constants';
 import { debounce } from '../../common/misc';
-import { type Environment, type EnvironmentKvPairData, EnvironmentType, getDataFromKVPair, getKVPairFromData } from '../../models/environment';
+import { type Environment, type EnvironmentKvPairData, EnvironmentType, getDataFromKVPair } from '../../models/environment';
 import { isRemoteProject } from '../../models/project';
 import { WorkspaceDropdown } from '../components/dropdowns/workspace-dropdown';
 import { WorkspaceSyncDropdown } from '../components/dropdowns/workspace-sync-dropdown';
 import { EditableInput } from '../components/editable-input';
 import { EnvironmentEditor, type EnvironmentEditorHandle, type EnvironmentInfo } from '../components/editors/environment-editor';
 import { EnvironmentKVEditor } from '../components/editors/environment-key-value-editor/key-value-editor';
+import { handleToggleEnvironmentType } from '../components/editors/environment-utils';
 import { Icon } from '../components/icon';
 import { useDocBodyKeyboardShortcuts } from '../components/keydown-binder';
-import { showModal } from '../components/modals';
 import { showAlert } from '../components/modals';
-import { AskModal } from '../components/modals/ask-modal';
 import { useOrganizationPermissions } from '../hooks/use-organization-features';
 import type { WorkspaceLoaderData } from './workspace';
 
@@ -459,14 +458,7 @@ const Environments = () => {
             {selectedEnvironment && (
               <ToggleButton
                 onChange={isSelected => {
-                  const newEnvironmentType = isSelected ? EnvironmentType.JSON : EnvironmentType.KVPAIR;
-                  // clear kvPairData when switch to json view, otherwise convert json data to kvPairData
-                  const kvPairData = isSelected ? [] : getKVPairFromData(selectedEnvironment.data, selectedEnvironment.dataPropertyOrder);
-                  const foundDisabledItem = isSelected && selectedEnvironment.kvPairData?.some(pair => !pair.enabled);
-                  const foundDuplicateNameItem = isSelected && selectedEnvironment.kvPairData?.some(
-                    (pair, idx) => selectedEnvironment.kvPairData?.slice(idx + 1).some(newPair => pair.name.trim() === newPair.name.trim() && newPair.enabled)
-                  );
-                  const toggleSwitchEnvironmentType = () => {
+                  const toggleSwitchEnvironmentType = (newEnvironmentType: EnvironmentType, kvPairData: EnvironmentKvPairData[]) => {
                     updateEnvironmentFetcher.submit(JSON.stringify({
                       patch: {
                         environmentType: newEnvironmentType,
@@ -479,25 +471,7 @@ const Environments = () => {
                       encType: 'application/json',
                     });
                   };
-                  if (foundDisabledItem || foundDuplicateNameItem) {
-                    showModal(AskModal, {
-                      title: 'Change Environment Type',
-                      message: (
-                        <>
-                          {foundDisabledItem && <p>All disabled items will be lost.</p>}
-                          {foundDuplicateNameItem && <p>Items with same name will be lost except the last one.</p>}
-                          <p>Are you sure to continue?</p>
-                        </>
-                      ),
-                      onDone: async (saidYes: boolean) => {
-                        if (saidYes) {
-                          toggleSwitchEnvironmentType();
-                        }
-                      },
-                    });
-                  } else {
-                    toggleSwitchEnvironmentType();
-                  }
+                  handleToggleEnvironmentType(isSelected, selectedEnvironment, toggleSwitchEnvironmentType);
                 }}
                 isSelected={selectedEnvironment?.environmentType !== EnvironmentType.KVPAIR}
                 className="w-[14ch] flex flex-shrink-0 gap-2 items-center justify-start px-2 py-1 rounded-sm text-[--color-font] hover:bg-[--hl-xs] focus:ring-inset ring-1 ring-transparent focus:ring-[--hl-md] transition-colors text-sm"
