+++ packages/insomnia/src/ui/routes/request-group.tsx
@@ -1,6 +1,7 @@
 import { type ActionFunction, type LoaderFunction, redirect } from 'react-router-dom';
 
 import * as models from '../../models';
+import { EnvironmentType } from '../../models/environment';
 import type { RequestGroup } from '../../models/request-group';
 import type { RequestGroupMeta } from '../../models/request-group-meta';
 import { invariant } from '../../utils/invariant';
@@ -28,7 +29,9 @@ export const createRequestGroupAction: ActionFunction = async ({ request, params
   const formData = await request.formData();
   const name = formData.get('name') as string;
   const parentId = formData.get('parentId') as string;
-  const requestGroup = await models.requestGroup.create({ parentId: parentId || workspaceId, name });
+  // New folder environment to be key-value pair by default;
+  const environmentType = formData.get('environmentType') as EnvironmentType || EnvironmentType.KVPAIR;
+  const requestGroup = await models.requestGroup.create({ parentId: parentId || workspaceId, name, environmentType });
   await models.requestGroupMeta.create({ parentId: requestGroup._id, collapsed: false });
   return null;
 };
